# Specialize sources for target and optimized kernels independently.