--- a/tensorflow/lite/micro/tools/make/ext_libs/xtensa_hifimini.inc
+++ b/tensorflow/lite/micro/tools/make/ext_libs/xtensa_hifimini.inc
@@ -1,0 +1,4 @@
+# Every optimized kernel implementation directory (i.e.
+# micro/kernels/<optimized_kernel_dir>/ must have a corresponding
+# micro/tools/make/ext_libs/<optimized_kernel_dir>.inc
+