This change:
 * adds a new command line option to the makefile called OPTIMIZED_KERNEL_DIR which should be used instead of TAGS

 * TAGS is currently supported for backwards compatibility and a gradual roll-out but will be dropped (for optimized kernels) in follow-up changes.

 * Makes the necessary change to use OPTIMIZED_KERNEL_DIR for xtensa_hifimini.

First step towards #44566
Fixes internal bug: http://b/160795217
Addresses internal bug: http://b/170501366

Tested the keyword_benchmark with different combinations command line options.

We used this command
```
make -f tensorflow/lite/micro/tools/make/Makefile -j8 TARGET=xtensa_hifimini  XTENSA_CORE=<core> test_keyword_benchmark
```
along with these additional options:
|   options                | InitializeKeywordRunner | KeywordRunNIerations(1) | KeywordRunNIerations(10) |
|        ------------------ | ---------------------- | ------------------------- | ------------------------- |
(on master) <br> `TAGS=xtensa_hifimini`        | 1389000 | 89318 | 892739 |
`TAGS=xtensa_hifimini`                                    | 1389056 | 89318 | 892739 |
`OPTIMIZED_KERNEL_DIR=xtensa_hifimini` | 1389056 | 89318 | 892739 |
`OPTIMIZED_KERNEL_DIR=none`                 |  226514| 1010031 | 10088115 |

The table is showing that:
 * we support `TAGS=xtensa_hifimini` as well as `OPTIMIZED_KERNEL_DIR=xtensa_hifimini` to pull in the optimized implementations.
 * `OPTIMIZED_KERNEL_DIR=none` uses the reference kernels.
 * There is a difference of 56 cycles in the initialization with and without this change (first two lines of the table). This can be chalked up to some small difference in how a potentially difference sequence of object files is handled.

We also see a small difference in the binary size:
```
 xt-size tensorflow/lite/micro/tools/make/gen/xtensa_hifimini_xtensa_hifimini/bin/keyword_benchmark
```
On master:
```
   text	   data	    bss	    dec
  54864	  48680	  25032	 128576
```

With this change (and either `TAGS=xtensa_hifimini` or `OPTIMIZED_KERNEL_DIR=xtensa_hifimini`):
```
   text	   data	    bss	    dec
  54864	  48664	  25032	 128560	
```